Change from VariantStructure to Entity
As I have removed the derivation, the visibility of Aspect disappeared, in
fact I don't even need it. Entities send each other messages, the logics
are linked to a subset of data only internally.
Changed the DustAPI to DustWorld, explanation in the ideas.txt.
The sandbox is a bit refactored for later extensions, still working with
BootWorld now.