Combined: Variant to interface, fieldType into Types, booting type management
The FieldType enum was misplaced (it is related to type management), so
it was put to its place. To do this, the DustVariant had to be converted
to interface, and its implementation went into DustUtilVariant, used by the
boot (and combined with the previous, removed DustBootVariant).

From now on, the requestor cannot tell the Variant type - it has to know
it and request the data accordingly. The handler logic is the kernel with
the type management information at runtime, while during the boot, the
DustUtilVariant automatic behaviors.

Using these features, I have created the boot object structures for type
management booting (this code will be generated by dustLab later on).