Database dump
Jars project created for external binaries; MySQL JDBC connector included
The generic (and a bit ugly) value extractor function was taken out of the
StreamDumper and reused in DBDumper.
The database looks good, it seems to be a correct dump.

Heading for DB serialization and read the content back from the DB.