DB Read
The Variant, Util, Boot components were extended for "lazy creation":
make an empty entity/aspect/variant, and fill their content later.

The Boot component was hacked to handle "Creating" and "Steady" state,
Creating entities are skipped from checking.

This all were necessare because in the Test I read back the entity describing
the Field type (thus there were multiple Field type descriptions in the
world).

The test looks fine for the first glance: the content of the Field entity
seems to look the same when dumping the boot and the db content.