DB Update
A bit dirty but quick solution: whe the entity has a persistentId, this
Id is used to identify its content, and during the save operation,
aspect content is deleted and regenerated (it was waay too much work to
select the old content, compare, update, etc.)

Context must be closely connected to this layer, a context commit should
mean all the local dirty content to be stored in the persistence component.

Internal entities (that can only live inside another) should also be handled
by the kernel: when their parent or link is deleted, they also should be
marked as deleted, and the commit should update the persistent store. It
is not the persistent layer's task to follow the aspect links and search
for internal entities.