Real source generation
Now the test program can generate valid Java sources in the proper
directory structure with the same content as the interfaces I create
manually for the types. The only difference is that I use different
prefix for the package name.
The type initialization is changed (units added to the types) in order
to have the correct package structure.

Now it is time to go on towards the GUI, then return back to the DB
persistence, to put the structure onto its own feet.