Towards sandbox dustLab (temporal app, later should be reimplemented)
The next task is to create a swing application that roughly looks like the
planned dustLab, to be able to manage the unit and type declarations.
Now the "left side" is under development: the list of current type and
unit entities on a list, doubleclick displays a JInternalFrame with the
selected entity, the selection is in sync between the list and the
desktop.

To come: entity editor panel, save, load, source generation.