JSON import works on syntax level
The problem root was that the processRelayReturn was not immediate
and recursive. After the refactor both the ProcessManager and the
template classes are nicer.
The source rollback went one character back because of the for cycle
step.
Some nice toString() functions added for easier debugging.

Next task: create the actual entities and aspects. Having it I can
import new types from JSON :-)