dust again...
I gave up MAF because of the ugly setData() functions, which are now
replaced with generated POJOs for the declared types. Successfully brought
type management kernel to life (basic types and dump).
I tried to replace the stream / indent functions with dust types, but hit
the wall again. I started with POJO function calls instead of send()
(following the bad memories with setData() ) - but for service calls send()
seems to be a better solution. The caller rather wants to control the message
processing (parallel call, fork, wait), and that is impossible with calling
generated (proxy) functions instead of sending messages.
The generated code helps the service writer who has to implement the interface
of the type, and a generated dispatcher transfers the message to the
actual function.