entity based dump
Introducing the first actual service components: the StreamWriter and Indenter.
They might look weird, but now the dump runs with them, so the stream and indent
was removed from the dust API const declarations.

Overload mechanism (aspect version of class inheritance)
 q'n'd implementation and usage: the Indenter overloads the streamwriter
interface, so it can follow its operation (to put the indent lead string
to the start of the line only when needed)